# docs/api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: TRM API
  description: |
    The Referral Marketplace (TRM) API allows developers to build on the TRM platform.
    Manage job postings, referrals, companies, and receive real-time webhook notifications.
  version: 1.0.0
  contact:
    name: TRM API Support
    email: api-support@trm.com
    url: https://docs.trm.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.trm.com/v1
    description: Production server
  - url: https://api.staging.trm.com/v1
    description: Staging server
security:
  - ApiKeyAuth: []
...
paths:
...
  /companies:
    get:
      summary: List companies
      description: Retrieve a list of companies
      tags:
        - Companies
      responses:
        '200':
          description: Successful response
    post:
      summary: Create company
      description: Create a new company (requires companies:write permission)
      tags:
        - Companies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyCreateRequest'
      responses:
        '201':
          description: Company created successfully
  /companies/{id}:
    get:
      summary: Get company
      description: Get detailed information about a company
      tags:
        - Companies
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
    put:
      summary: Update company
      description: Update a company (requires companies:write permission)
      tags:
        - Companies
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Company updated successfully
  /companies/{id}/jobs:
    get:
      summary: Get company jobs
      description: Get all jobs for a specific company
      tags:
        - Companies
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication
  schemas:
...
    CompanyCreateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
        website:
          type: string
        industry:
          type: string
...